# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carpool Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/rides/my-rides": {
            "get": {
                "tags": [
                    "Ride"
                ],
                "summary": "get list of my rides",
                "description": "get list of my rides",
                "operationId": "getMyRides",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/rides": {
            "get": {
                "tags": [
                    "Ride"
                ],
                "summary": "get list of rides",
                "description": "get list of rides",
                "operationId": "getAllRides",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filters[origin]",
                        "in": "query",
                        "description": "origin",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "filters[destination]",
                        "in": "query",
                        "description": "destination",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "filters[date]",
                        "in": "query",
                        "description": "date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2023-11-22"
                        }
                    },
                    {
                        "name": "filters[capacity]",
                        "in": "query",
                        "description": "capacity",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "filters[gender]",
                        "in": "query",
                        "description": "gender",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer token for Sanctum authentication"
            }
        }
    ]
}